nvif: split out fermi interface definitions
